core/serializers.py
from rest_framework import serializers
from .models import File


class ChangeEmailSerializer(serializers.Serializer):
    email = serializers.EmailField()
    confirm_email = serializers.EmailField()


class ChangePasswordSerializer(serializers.Serializer):
    password = serializers.CharField(style={'input_type': 'password'})
    confirm_password = serializers.CharField(style={'input_type': 'password'})
    current_password = serializers.CharField(style={'input_type': 'password'})


class FileSerializer(serializers.ModelSerializer):
    class Meta:
        model = File
        fields = "__all__"
