requirements.txt
autopep8==1.4.4
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.2.1
django-allauth==0.37.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
gunicorn==19.7.1
idna==2.7
numpy==1.16.3
oauthlib==2.1.0
opencv-python==4.1.0.25
pep8==1.7.1
Pillow==6.0.0
pycodestyle==2.5.0
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.3.0
stripe==2.28.1
urllib3==1.24.3
whitenoise==3.3.1
